Optimize Trim() and move helpers to SpanHelpers.char.cs
